Add about section image and proper flex layout
Took some extra time to make sure that the about section has a proper
flex layout as it was a bit freeform and explicit. This is a bit easier
to manipulate in the future. Added the image to the about section and
made sure it is circular and properly formatted.